+++ src/main/java/by/shevchenko/graduationproject/config/SecurityConfig.java
@@ -24,10 +24,10 @@ public class SecurityConfig {
                 .csrf().disable()
                 .cors().disable()
                 .authorizeRequests((auth) -> auth
-                        .requestMatchers("/registration", "/login", "/swagger-ui/**", "/v3/api-docs/**", "/product", "/product/")
+                        .requestMatchers("/registration", "/login", "/swagger-ui/**", "/v3/api-docs/**", "/product/**")
                         .permitAll()
-                        .requestMatchers("/product/**").hasRole("ADMIN")
-                        .requestMatchers("basket/**").hasRole("USER")
+                        .requestMatchers("/admin/**").hasRole("ADMIN")
+                        .requestMatchers("/basket/**").hasRole("USER")
                         .anyRequest()
                         .authenticated()
                         .and()
