+++ src/main/java/by/shevchenko/graduationproject/controller/ProductController.java
@@ -10,7 +10,6 @@ import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -24,25 +23,6 @@ import java.util.List;
 public class ProductController {
     private final ProductService productService;
 
-    @Operation(summary = "add product to date"
-            , description = "this method return ProductEntity"
-            , responses = @ApiResponse(responseCode = "200"
-            , description = "ProductEntity"
-            , content = {@Content(mediaType = "application/json"
-            , schema = @Schema(implementation = ProductEntity.class)
-            , examples = @ExampleObject(value = "{\n" +
-            "  \"id\": 1,\n" +
-            "  \"name\": \"Iphone 13ProMax\",\n" +
-            "  \"inStock\": true,\n" +
-            "  \"quantity\": 2,\n" +
-            "  \"price\": 1000\n" +
-            "}"))}))
-
-    @ResponseStatus(HttpStatus.OK)
-    @PostMapping("{id}")
-    public ProductEntity addProduct(@RequestBody ProductEntity product) {
-        return productService.add(product);
-    }
 
     @Operation(summary = "find by id product"
             , description = "this method return ProductEntity"
@@ -92,36 +72,5 @@ public class ProductController {
         return productService.findAll();
     }
 
-    @Operation(summary = "update product"
-            , description = "this method return ProductEntity"
-            , responses = {@ApiResponse(responseCode = "200"
-            , description = "ProductEntity"
-            , content = {@Content(mediaType = "application/json"
-            , schema = @Schema(implementation = ProductEntity.class)
-            , examples = @ExampleObject(value = "{\n" +
-            "  \"id\": 1,\n" +
-            "  \"name\": \"Iphone 12\",\n" +
-            "  \"inStock\": true,\n" +
-            "  \"quantity\": 2,\n" +
-            "  \"price\": 1000\n" +
-            "}"))})
-            , @ApiResponse(responseCode = "404"
-            , description = "Product not find"
-            , content = @Content)})
-    @GetMapping("update/{productId}/{quantity}")
-    public ProductEntity update(@PathVariable Long productId, @PathVariable int quantity) {
-        return productService.update(productId, quantity);
-    }
 
-    @Operation(summary = "Removes a product from date"
-            , description = "method returns nothing"
-            , responses = {@ApiResponse(responseCode = "200")
-            , @ApiResponse(responseCode = "404"
-            , description = "Product not find"
-            , content = @Content)})
-    @ResponseStatus(HttpStatus.OK)
-    @DeleteMapping("{id}")
-    public void delete(@PathVariable Long id) {
-        productService.delete(id);
-    }
 }
